requirements.txt
psycopg==3.2.3
Django==5.1.1